+++ src/world/mod.rs
@@ -13,7 +13,7 @@ struct World {
     size: CoordinateTriplet<f64>,
 
     /// number of cells
-    cells: CoordinateTriplet<u64>,
+    cells: CoordinateTriplet<usize>,
 
     /// (m) spatial increment
     delta: CoordinateTriplet<f64>,
@@ -24,19 +24,19 @@ impl World {
     ///
     /// # Arguments
     /// - `size`: &[f64; 3] (m) size of bounding box
-    /// - `cells`: &[u64; 3] number of cells
+    /// - `cells`: &[usize; 3] number of cells
     ///
     /// # Returns
     /// `Result<World, anyhow::Error>`
     ///
     /// # Errors
     /// - any call to `CoordinateTriplet::new()` fails
-    pub fn new(size: &[f64; 3], cells: &[u64; 3]) -> Result<World, anyhow::Error> {
+    pub fn new(size: &[f64; 3], cells: &[usize; 3]) -> Result<World, anyhow::Error> {
         // unpack dimensions
         let size: CoordinateTriplet<f64> = CoordinateTriplet::new(size[0], size[1], size[2])?;
 
         // unpack cells
-        let cells: CoordinateTriplet<u64> = CoordinateTriplet::new(cells[0], cells[1], cells[2])?;
+        let cells: CoordinateTriplet<usize> = CoordinateTriplet::new(cells[0], cells[1], cells[2])?;
 
         // initialize spatial increments
         let dx = size.x / (cells.x - 1) as f64;
@@ -65,7 +65,7 @@ mod tests {
     ///
     fn setup() -> Result<World, anyhow::Error> {
         let size: [f64; 3] = [1.0, 2.0, 3.0];
-        let cells: [u64; 3] = [3, 11, 31];
+        let cells: [usize; 3] = [3, 11, 31];
         World::new(&size, &cells)
     }
